Redo header index

+++ src/Chain/CachingUtxoSet.php
@@ -2,12 +2,13 @@
 
 namespace BitWasp\Bitcoin\Node\Chain;
 
-use BitWasp\Bitcoin\Node\DbInterface;
+use BitWasp\Bitcoin\Node\Db\DbInterface;
 use BitWasp\Bitcoin\Script\Script;
 use BitWasp\Bitcoin\Serializer\Transaction\OutPointSerializer;
 use BitWasp\Bitcoin\Transaction\OutPointInterface;
 use BitWasp\Bitcoin\Transaction\TransactionOutput;
 use BitWasp\Bitcoin\Utxo\Utxo;
+use BitWasp\Bitcoin\Utxo\UtxoInterface;
 use BitWasp\Buffertools\Buffer;
 use Doctrine\Common\Cache\ArrayCache;
 use Doctrine\Common\Cache\RedisCache;
@@ -20,10 +21,18 @@ class CachingUtxoSet
     private $db;
 
     /**
+     * @var bool
+     */
+    private $caching = false;
+
+    /**
      * @var RedisCache
      */
     private $set;
 
+    /**
+     * @var array
+     */
     private $cacheHits = [];
 
     /**
@@ -57,35 +66,30 @@ class CachingUtxoSet
      */
     public function applyBlock(array $deleteOutPoints, array $newUtxos)
     {
-        $this->db->transaction(function () use ($deleteOutPoints, $newUtxos) {
-            //if (!empty($this->cacheHits)) {
-//                $this->db->appendUtxoViewKeys($this->cacheHits);
-//            }
+        $this->db->updateUtxoSet($this->outpointSerializer, $deleteOutPoints, $newUtxos, $this->cacheHits);
+        
+        if ($this->caching) {
+            foreach ($this->cacheHits as $key) {
+                $this->set->delete($key);
+            }
 
-            $this->db->updateUtxoSet($this->outpointSerializer, $deleteOutPoints, $newUtxos, $this->cacheHits);
+            foreach ($newUtxos as $c => $utxo) {
+                $new = $this->outpointSerializer->serialize($utxo->getOutPoint())->getBinary();
+                $this->set->save($new, [
+                    $newUtxos[$c]->getOutput()-> getValue(),
+                    $newUtxos[$c]->getOutput()->getScript()->getBinary(),
+                ], 500000);
+            }
 
-        });
+            echo "Inserts: " . count($newUtxos). " | Deletes: " . count($deleteOutPoints). " | " . "CacheHits: " . count($this->cacheHits) .PHP_EOL;
 
-        foreach ($this->cacheHits as $key) {
-            $this->set->delete($key);
+            $this->cacheHits = [];
         }
-
-        foreach ($newUtxos as $c => $utxo) {
-            $new = $this->outpointSerializer->serialize($utxo->getOutPoint())->getBinary();
-            $this->set->save($new, [
-                $newUtxos[$c]->getOutput()-> getValue(),
-                $newUtxos[$c]->getOutput()->getScript()->getBinary(),
-            ], 500000);
-        }
-
-        echo "Inserts: " . count($newUtxos). " | Deletes: " . count($deleteOutPoints). " | " . "CacheHits: " . count($this->cacheHits) .PHP_EOL;
-
-        $this->cacheHits = [];
     }
 
     /**
      * @param OutPointInterface[] $requiredOutpoints
-     * @return UtxoView
+     * @return UtxoInterface[]
      */
     public function fetchView(array $requiredOutpoints)
     {
@@ -93,18 +97,14 @@ class CachingUtxoSet
             $utxos = [];
             $required = [];
             $cacheHits = [];
-            $a = 0;
-            $b = 0;
             foreach ($requiredOutpoints as $c => $outpoint) {
                 $key = $this->outpointSerializer->serialize($outpoint)->getBinary();
                 if ($this->set->contains($key)) {
                     list ($value, $scriptPubKey) = $this->set->fetch($key);
                     $cacheHits[] = $key;
                     $utxos[] = new Utxo($outpoint, new TransactionOutput($value, new Script(new Buffer($scriptPubKey))));
-                    $a++;
                 } else {
                     $required[] = $outpoint;
-                    $b++;
                 }
             }
 
@@ -112,10 +112,13 @@ class CachingUtxoSet
                 $utxos = array_merge($utxos, $this->db->fetchUtxoDbList($this->outpointSerializer, $required));
             }
 
-            $this->cacheHits = $cacheHits;
+            if ($this->caching) {
+                $this->cacheHits = $cacheHits;
+            }
 
             return $utxos;
         } catch (\Exception $e) {
+            echo $e->getMessage().PHP_EOL;
             throw new \RuntimeException('Failed to find UTXOS in set');
         }
     }
